# api/api.spec.json
{
	"swagger": "2.0",
	"info": {
	  "version": "0.0.1",
	  "title": "Twitch Watcher API",
	  "description": "The API that will describe the twitch watching service",
	  "contact": {
		"name": "Evan Coulson"
	  }
	},
...
	"definitions": {
	  "TwitchUserWebhookInfo": {
		"properties": {
		  "id": {
			"type": "integer"
		  },
		  "streamHook": {
			"$ref": "#/definitions/WebhookInfo"
		  },
		  "userHook": {
			"$ref": "#/definitions/WebhookInfo"
		  },
		  "followerHook": {
			"$ref": "#/definitions/WebhookInfo"
		  }
		}
	  },
	  "WebhookInfo": {
		"properties": {
		  "callbackURL": {
			"type": "string"
		  },
		  "topicURL": {
			"type": "string"
		  },
		  "expirationDate": {
			"type": "string",
			"format": "date-time"
		  }
		}
	  },
	  "Message": {
		"properties": {
		  "message": {
			"type": "string"
		  },
		  "success": {
			"type": "boolean"
		  }
		}
	  },
	  "SubscribeBody": {
		"type": "object",
		"properties": {
		  "callback": {
			"type": "string"
		  },
		  "userID": {
			"type": "integer"
		  }
		}
	  },
	  "UnsubscribeBody": {
		"type": "object",
		"properties": {
		  "userID": {
			"type": "integer"
		  }
		}
	  },
...
	},
	"paths": {
...
	  "/subscribe": {
		"post": {
		  "description": "The endpoint that handles subscriptions to a Twitch webhook topic.",
		  "summary": "Initiates a Twitch web hook subscription to a given topic",
		  "produces": [
			"application/json"
		  ],
		  "parameters": [
			{
			  "in": "body",
			  "name": "body",
			  "description": "The body of the subscription request",
			  "required": true,
			  "schema": {
				"$ref": "#/definitions/SubscribeBody"
			  }
			}
		  ],
		  "operationId": "subscribeToHook",
		  "responses": {
			"200": {
			  "description": "Response from a successful subscripition",
			  "schema": {
				"$ref": "#/definitions/TwitchUserWebhookInfo"
			  }
			},
			"400": {
			  "description": "Response from a failed subscripition",
			  "schema": {
				"$ref": "#/definitions/Message"
			  }
			}
		  },
		  "tags": [
			"Subscriptions",
			"User"
		  ]
		}
	  },
	  "/unsubscribe": {
		"post": {
		  "description": "The endpoint that handles unsubscription from a Twitch webhook topic.",
		  "summary": "Initiates a Twitch web hook subscription to a given topic",
		  "produces": [
			"application/json"
		  ],
		  "parameters": [
			{
			  "in": "body",
			  "name": "body",
			  "description": "The body of the subscription request",
			  "required": true,
			  "schema": {
				"$ref": "#/definitions/UnsubscribeBody"
			  }
			}
		  ],
		  "operationId": "unsubscribeFromHook",
		  "responses": {
			"200": {
			  "description": "Response from a successful unsubscripition",
			  "schema": {
				"$ref": "#/definitions/TwitchUserWebhookInfo"
			  }
			},
			"400": {
			  "description": "Response from a failed unsubscripition",
			  "schema": {
				"$ref": "#/definitions/Message"
			  }
			}
		  },
		  "tags": [
			"Subscriptions",
			"User"
		  ]
		}
	  },
...
	}
  }